+++ R/reporter-check.R
@@ -11,41 +11,44 @@ NULL
 CheckReporter <- R6::R6Class("CheckReporter",
   inherit = Reporter,
   public = list(
-    failures = list(),
+    problems = NULL,
+    skips = NULL,
+
     n_ok = 0L,
     n_skip = 0L,
     n_fail = 0L,
     n_warn = 0L,
 
-    skips = NULL,
-
-    stop_on_failure = TRUE,
-
-    initialize = function(stop_on_failure = TRUE, ...) {
-      self$stop_on_failure <- stop_on_failure
+    initialize = function(...) {
+      self$problems <- Stack$new()
       self$skips <- Stack$new()
+
       super$initialize(...)
     },
 
     add_result = function(context, test, result) {
-      if (expectation_skip(result)) {
-        self$n_skip <- self$n_skip + 1L
-        self$skips$push(result$message)
-        return()
-      }
-      if (expectation_warning(result)) {
-        self$n_warn <- self$n_warn + 1L
-        return()
-      }
-      if (expectation_ok(result)) {
+      if (expectation_success(result)) {
         self$n_ok <- self$n_ok + 1L
         return()
       }
 
-      self$n_fail <- self$n_fail + 1L
-      self$failures[[self$n_fail]] <- result
+      self$problems$push(result)
 
-      self$cat_line(failure_summary(result, self$n_fail))
+      if (expectation_skip(result)) {
+        self$n_skip <- self$n_skip + 1L
+        self$skips$push(result$message)
+      } else if (expectation_warning(result)) {
+        self$n_warn <- self$n_warn + 1L
+      } else {
+        self$n_fail <- self$n_fail + 1L
+      }
+
+      type <- expectation_type(result)
+      header <- failure_header(result)
+
+      self$local_user_output()
+      self$rule(header, col = testthat_style(type))
+      self$cat_line(format(result, simplify = "none"))
       self$cat_line()
     },
 
@@ -66,60 +69,31 @@ CheckReporter <- R6::R6Class("CheckReporter",
         " ]"
       )
 
-      if (self$n_fail == 0) return()
+      problems <- self$problems$as_list()
+      if (length(problems) == 0) {
+        return()
+      }
+      saveRDS(problems, "testthat-problems.rds")
 
-      if (self$n_fail > 10) {
-        show <- self$failures[1:9]
+      if (length(problems) > 10) {
+        show <- problems[1:9]
       } else {
-        show <- self$failures
+        show <- problems
       }
 
       fails <- vapply(show, failure_header, character(1))
-      if (self$n_fail > 10) {
-        fails <- c(fails, "...")
-      }
       labels <- format(paste0(1:length(show), "."))
       self$cat_line(paste0(labels, " ", fails, collapse = "\n"))
-      self$cat_line()
-
-      if (self$stop_on_failure) {
-        stop("testthat unit tests failed", call. = FALSE)
+      if (length(problems) > 10) {
+        self$cat_line("... and ", length(problems) - 10, " more")
       }
+      self$cat_line()
     }
   )
 )
 
-
-skip_summary <- function(x, label) {
-  header <- paste0(label, ". ", x$test)
-
-  paste0(
-    colourise(header, "skip"), src_loc(x$srcref), " - ", x$message
-  )
-}
-
-failure_summary <- function(x, label, width = cli::console_width()) {
-  header <- paste0(label, ". ", failure_header(x))
-
-  paste0(
-    cli::rule(header, col = testthat_style("error")), "\n",
-    format(x)
-  )
-}
-
 failure_header <- function(x) {
-  type <- switch(expectation_type(x),
-    error = "Error",
-    failure = "Failure"
-  )
-
-  paste0(type, ": ", x$test, src_loc(x$srcref), " ")
-}
-
-src_loc <- function(ref) {
-  if (is.null(ref)) {
-    ""
-  } else {
-    paste0(" (@", basename(attr(ref, "srcfile")$filename), "#", ref[1], ")")
-  }
+  type <- expectation_type(x)
+  substr(type, 1, 1) <- toupper(substr(type, 1, 1))
+  paste0(type, ": ", x$test, " (", expectation_location(x), ")")
 }